[代码完善](v2.5): v2.5 beta mapper包 改为 mapstruct
2.5 Beta 详情：https://www.ydyno.com/archives/1225.html
