+++ eladmin-logging/src/main/java/me/zhengjie/service/impl/LogServiceImpl.java
@@ -23,8 +23,8 @@ import me.zhengjie.domain.Log;
 import me.zhengjie.repository.LogRepository;
 import me.zhengjie.service.LogService;
 import me.zhengjie.service.dto.LogQueryCriteria;
-import me.zhengjie.service.mapper.LogErrorMapper;
-import me.zhengjie.service.mapper.LogSmallMapper;
+import me.zhengjie.service.mapstruct.LogErrorMapper;
+import me.zhengjie.service.mapstruct.LogSmallMapper;
 import me.zhengjie.utils.*;
 import org.aspectj.lang.ProceedingJoinPoint;
 import org.aspectj.lang.reflect.MethodSignature;
