+++ eladmin-system/src/main/java/me/zhengjie/modules/mnt/service/mapstruct/DatabaseMapper.java
@@ -13,19 +13,19 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package me.zhengjie.modules.system.service.mapper;
+package me.zhengjie.modules.mnt.service.mapstruct;
 
 import me.zhengjie.base.BaseMapper;
-import me.zhengjie.modules.system.domain.Menu;
-import me.zhengjie.modules.system.service.dto.MenuDto;
+import me.zhengjie.modules.mnt.domain.Database;
+import me.zhengjie.modules.mnt.service.dto.DatabaseDto;
 import org.mapstruct.Mapper;
 import org.mapstruct.ReportingPolicy;
 
 /**
- * @author Zheng Jie
- * @date 2018-12-17
- */
+* @author zhanghouying
+* @date 2019-08-24
+*/
 @Mapper(componentModel = "spring",unmappedTargetPolicy = ReportingPolicy.IGNORE)
-public interface MenuMapper extends BaseMapper<MenuDto, Menu> {
+public interface DatabaseMapper extends BaseMapper<DatabaseDto, Database> {
 
 }
