+++ eladmin-system/src/main/java/me/zhengjie/modules/mnt/service/mapstruct/DeployMapper.java
@@ -13,11 +13,11 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package me.zhengjie.modules.mnt.service.mapper;
+package me.zhengjie.modules.mnt.service.mapstruct;
 
 import me.zhengjie.base.BaseMapper;
-import me.zhengjie.modules.mnt.domain.DeployHistory;
-import me.zhengjie.modules.mnt.service.dto.DeployHistoryDto;
+import me.zhengjie.modules.mnt.domain.Deploy;
+import me.zhengjie.modules.mnt.service.dto.DeployDto;
 import org.mapstruct.Mapper;
 import org.mapstruct.ReportingPolicy;
 
@@ -25,7 +25,7 @@ import org.mapstruct.ReportingPolicy;
 * @author zhanghouying
 * @date 2019-08-24
 */
-@Mapper(componentModel = "spring",uses = {},unmappedTargetPolicy = ReportingPolicy.IGNORE)
-public interface DeployHistoryMapper extends BaseMapper<DeployHistoryDto, DeployHistory> {
+@Mapper(componentModel = "spring",uses = {AppMapper.class, ServerDeployMapper.class},unmappedTargetPolicy = ReportingPolicy.IGNORE)
+public interface DeployMapper extends BaseMapper<DeployDto, Deploy> {
 
 }
