+++ eladmin-system/src/main/java/me/zhengjie/modules/system/service/mapstruct/JobMapper.java
@@ -13,19 +13,18 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package me.zhengjie.modules.system.service.mapper;
+package me.zhengjie.modules.system.service.mapstruct;
 
 import me.zhengjie.base.BaseMapper;
-import me.zhengjie.modules.system.domain.Dept;
-import me.zhengjie.modules.system.service.dto.DeptDto;
+import me.zhengjie.modules.system.domain.Job;
+import me.zhengjie.modules.system.service.dto.JobDto;
 import org.mapstruct.Mapper;
 import org.mapstruct.ReportingPolicy;
 
 /**
 * @author Zheng Jie
-* @date 2019-03-25
+* @date 2019-03-29
 */
-@Mapper(componentModel = "spring",unmappedTargetPolicy = ReportingPolicy.IGNORE)
-public interface DeptMapper extends BaseMapper<DeptDto, Dept> {
-
+@Mapper(componentModel = "spring",uses = {DeptMapper.class},unmappedTargetPolicy = ReportingPolicy.IGNORE)
+public interface JobMapper extends BaseMapper<JobDto, Job> {
 }
