+++ eladmin-system/src/main/java/me/zhengjie/modules/system/service/mapstruct/MenuMapper.java
@@ -13,19 +13,19 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package me.zhengjie.modules.system.service.mapper;
+package me.zhengjie.modules.system.service.mapstruct;
 
 import me.zhengjie.base.BaseMapper;
-import me.zhengjie.modules.system.domain.Dept;
-import me.zhengjie.modules.system.service.dto.DeptSmallDto;
+import me.zhengjie.modules.system.domain.Menu;
+import me.zhengjie.modules.system.service.dto.MenuDto;
 import org.mapstruct.Mapper;
 import org.mapstruct.ReportingPolicy;
 
 /**
-* @author Zheng Jie
-* @date 2019-03-25
-*/
+ * @author Zheng Jie
+ * @date 2018-12-17
+ */
 @Mapper(componentModel = "spring",unmappedTargetPolicy = ReportingPolicy.IGNORE)
-public interface DeptSmallMapper extends BaseMapper<DeptSmallDto, Dept> {
+public interface MenuMapper extends BaseMapper<MenuDto, Menu> {
 
 }
