+++ eladmin-system/src/main/java/me/zhengjie/modules/system/service/mapstruct/UserMapper.java
@@ -13,19 +13,18 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package me.zhengjie.service.mapper;
+package me.zhengjie.modules.system.service.mapstruct;
 
 import me.zhengjie.base.BaseMapper;
-import me.zhengjie.domain.Log;
-import me.zhengjie.service.dto.LogErrorDTO;
+import me.zhengjie.modules.system.domain.User;
+import me.zhengjie.modules.system.service.dto.UserDto;
 import org.mapstruct.Mapper;
 import org.mapstruct.ReportingPolicy;
 
 /**
  * @author Zheng Jie
- * @date 2019-5-22
+ * @date 2018-11-23
  */
-@Mapper(componentModel = "spring",unmappedTargetPolicy = ReportingPolicy.IGNORE)
-public interface LogErrorMapper extends BaseMapper<LogErrorDTO, Log> {
-
+@Mapper(componentModel = "spring",uses = {RoleMapper.class, DeptMapper.class, JobMapper.class},unmappedTargetPolicy = ReportingPolicy.IGNORE)
+public interface UserMapper extends BaseMapper<UserDto, User> {
 }
