+++ eladmin-system/src/main/resources/template/generator/admin/ServiceImpl.ftl
@@ -32,7 +32,7 @@ import ${package}.repository.${className}Repository;
 import ${package}.service.${className}Service;
 import ${package}.service.dto.${className}Dto;
 import ${package}.service.dto.${className}QueryCriteria;
-import ${package}.service.mapper.${className}Mapper;
+import ${package}.service.mapstruct.${className}Mapper;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
